Refactor squad management functionality
This commit introduces significant refactoring of the squad management features in the BowlingMegabucks Tournament Manager application. Key updates include:

- Enhanced documentation with XML comments for better readability and maintainability.
- Improved handling of squad addition in `AddSquadAdapter` and `AddSquadPresenter`, including error management and asynchronous operations.
- Introduction of `CompleteSquadAdapter` and `CompleteSquadPresenter` for managing squad completion with error handling and user confirmation.
- Enhancements to `SquadResultsAdapter` and `SquadResultsPresenter` for retrieving and displaying squad results with improved error handling.
- Updates to `RetrieveSquadsAdapter` and `RetrieveSquadsPresenter` for squad retrieval, including adding new squads and refreshing the list.
- Modifications to `SquadViewModel` and `SquadResultsViewModel` for better data representation and conversion between view models and domain models.

These changes improve the structure, clarity, and functionality of the squad management features, making the codebase more robust and easier to work with.
