--- source/main/Squads/Retrieve/IRetrieveSquadsView.cs
@@ -1,16 +1,0 @@
-﻿namespace BowlingMegabucks.TournamentManager.Squads.Retrieve;
-
-internal interface IView
-{
-    TournamentId TournamentId { get; }
-
-    void BindSquads(IEnumerable<IViewModel> squads);
-
-    void Disable();
-
-    void DisplayError(string message);
-
-    SquadId? AddSquad(TournamentId tournamentId);
-
-    Task RefreshSquadsAsync(CancellationToken cancellationToken);
-}
