--- source/main/Squads/Retrieve/RetrieveSquadsPresenter.cs
@@ -1,52 +1,0 @@
-﻿using Microsoft.Extensions.DependencyInjection;
-
-namespace BowlingMegabucks.TournamentManager.Squads.Retrieve;
-internal class Presenter
-{
-    private readonly IView _view;
-
-    private readonly IAdapter _getSquadsAdapter;
-
-    public Presenter(IView view, IServiceProvider services)
-    {
-        _view = view;
-
-        _getSquadsAdapter = services.GetRequiredService<IAdapter>();
-    }
-
-    /// <summary>
-    /// Unit Test Constructor
-    /// </summary>
-    /// <param name="mockView"></param>
-    /// <param name="mockGetSquadsAdapter"></param>
-    internal Presenter(IView mockView, IAdapter mockGetSquadsAdapter)
-    {
-        _view = mockView;
-        _getSquadsAdapter = mockGetSquadsAdapter;
-    }
-
-    public async Task ExecuteAsync(CancellationToken cancellationToken)
-    {
-        var squads = await _getSquadsAdapter.ExecuteAsync(_view.TournamentId, cancellationToken).ConfigureAwait(true);
-
-        if (_getSquadsAdapter.Error != null)
-        {
-            _view.Disable();
-            _view.DisplayError(_getSquadsAdapter.Error.Message);
-        }
-        else
-        {
-            _view.BindSquads(squads.OrderBy(squad => squad.Date));
-        }
-    }
-
-    internal async Task AddSquadAsync(CancellationToken cancellationToken)
-    {
-        var squadId = _view.AddSquad(_view.TournamentId);
-
-        if (squadId != null)
-        {
-            await _view.RefreshSquadsAsync(cancellationToken).ConfigureAwait(true);
-        }
-    }
-}
