+++ source/presentation/BowlingMegabucks.TournamentManager.Presentation/PresentationExtensions.cs
@@ -3,6 +3,7 @@ using BowlingMegabucks.TournamentManager.Divisions;
 using BowlingMegabucks.TournamentManager.LaneAssignments;
 using BowlingMegabucks.TournamentManager.Registrations;
 using BowlingMegabucks.TournamentManager.Scores;
+using BowlingMegabucks.TournamentManager.Squads;
 using Microsoft.Extensions.Configuration;
 using Microsoft.Extensions.DependencyInjection;
 
@@ -26,7 +27,8 @@ public static class PresentationExtensions
             .AddDivisionModule()
             .AddLaneAssignmentsModule()
             .AddRegistrationsModule()
-            .AddScoresModule();
+            .AddScoresModule()
+            .AddSquadsModule();
 
         return services;
     }
