+++ source/presentation/BowlingMegabucks.TournamentManager.Presentation/Squads/Complete/CompleteSquadAdapter.cs
