+++ source/presentation/BowlingMegabucks.TournamentManager.Presentation/Squads/Results/SquadResultsAdapter.cs
